+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## main
+ - Add the "no_noise" option to the "windows" section of the config to allow for processing older waveforms that did not include pre-event noise. 
+ - Update workspace so that when the config attribute is set, any missing values entries are filled in with defaults.
+   
+
+
 ## 1.2.6 / 2023-03-18
  - Update URL in download code.
  - Bugfixes